Fixed the logic in the withCRS method so that it now returns a WarpRasterSource instead of an anonymous class
